# Iteration perf improvements