--- a/benchmark/issue25.js
+++ b/benchmark/issue25.js
@@ -14,9 +14,13 @@ function ForEachBench() {
   console.log("starting forEach benchmark");
   const tb = new TypedFastBitSet();
   const stb = new SparseTypedFastBitSet();
+  const tb_small = new TypedFastBitSet();
 
   const s = new Set();
   for (let i = 0; i < N; i++) {
+    if (i < 100) {
+      tb_small.add(i);
+    }
     tb.add(smallgap * i + 5);
     stb.add(smallgap * i + 5);
   }
@@ -24,36 +28,58 @@ function ForEachBench() {
   const suite = new Benchmark.Suite();
   // add tests
   const ms = suite
-  .add("TypedFastBitSet-forof", function () {
-    let card = 0;
-    for (const element of stb) {
-      card++;
-    }
-    return card;
-  })
-  .add("TypedFastBitSet-foreach", function () {
-    let card = 0;
-    const inc = function () {
-      card++;
-    };
-    tb.forEach(inc);
-    return card;
-  })
-  .add("SparseTypedFastBitSet-forof", function () {
-    let card = 0;
-    for (const element of stb) {
-      card++;
-    }
-    return card;
-  })
-  .add("SparseTypedFastBitSet-foreach", function () {
-    let card = 0;
-    const inc = function () {
-      card++;
-    };
-    tb.forEach(inc);
-    return card;
-  })
+    .add("TypedFastBitSet-forof", function () {
+      let card = 0;
+      for (const element of stb) {
+        card++;
+      }
+      return card;
+    })
+    .add("TypedFastBitSet-foreach", function () {
+      let card = 0;
+      const inc = function () {
+        card++;
+      };
+      tb.forEach(inc);
+      return card;
+    })
+    .add("SparseTypedFastBitSet-forof", function () {
+      let card = 0;
+      for (const element of stb) {
+        card++;
+      }
+      return card;
+    })
+    .add("SparseTypedFastBitSet-foreach", function () {
+      let card = 0;
+      const inc = function () {
+        card++;
+      };
+      tb.forEach(inc);
+      return card;
+    })
+    .add("TypedFastBitSet-nested-forof", function () {
+      let card = 0;
+      for (const x of tb_small) {
+        for (const y of tb_small) {
+          for (const z of tb_small) {
+            card += x + y + z;
+          }
+        }
+      }
+      return card;
+    })
+    .add("TypedFastBitSet-nested-forEach", function () {
+      let card = 0;
+      tb_small.forEach(function (x) {
+        tb_small.forEach(function (y) {
+          tb_small.forEach(function (z) {
+            card += x + y + z;
+          });
+        });
+      });
+      return card;
+    })
     // add listeners
     .on("cycle", function (event) {
       console.log(String(event.target));
@@ -61,7 +87,6 @@ function ForEachBench() {
     // run async
     .run({ async: false });
 }
-
 
 const main = function () {
   console.log(