Here are two commits that drastically improve the performance of using `TypedFastBitSet` and `SparseTypedFastBitSet` in `for..of` loops.

Just to explain: `*[Symbol.iterator]()` implementations, while terse, leverage JavaScript generators which are notably slower than just returning a plain iterator object. This is because generators also support coroutines, which aren't being leveraged at all here.

I'm making this PR because our team is favoring `forEach` over `for..of` for performance reasons, and if the point of this type is to make things fast, I think it should be fast in all cases. `forEach(fn)`, in particular when used in a nested fashion, don't perform as well as in-lined for loops. :) 

Thank you for this library BTW!