@lemire 2023-05-08
>  if the point of this type is to make things fast, I think it should be fast in all cases

I have checked in a new benchmark script at https://github.com/lemire/TypedFastBitSet.js/blob/master/benchmark/issue25.js

Could you run it?

I am getting that the for-of construction, with your PR, is still 2x slower than forEach.


```
TypedFastBitSet-forof x 121 ops/sec ±0.55% (79 runs sampled)
TypedFastBitSet-foreach x 248 ops/sec ±0.79% (92 runs sampled)
SparseTypedFastBitSet-forof x 123 ops/sec ±0.78% (80 runs sampled)
SparseTypedFastBitSet-foreach x 198 ops/sec ±0.28% (88 runs sampled)
```

I think we will merge your PR, but if your expectation is that you have bridged the performance gap... I am not sure that's the case. You should still use forEach if you care about performance, it seems.


Or do I misunderstand?