@benlesh 2023-05-08
:) Okay, so this explanation will get weird, so I'm including an update to your benchmark.

The implementation of `forEach` will ALWAYS be faster in the naive case for the exact same reason I'd like to make this change: Inlined loops are way faster than repeat function calls.

While the iterator might still be 2x slower than `forEach`, 2x slower than lightning fast is still better than 10x slower than lightning fast (which is what we had with the generator-based implementation)

And, when the only "fast" tool provided is `forEach`, that makes developers lean into nested `forEach` loops, which don't optimize well/naturally in the runtime. This is because the most obvious/readable implementations will involve inlined functions (usually arrow functions) that must be allocated and GC'ed during each loop. The deeper the nesting the worse it is.

I've included benchmark examples showing a nesting of `for..of` and `forEach` 3x deep over a smaller set of data. You'll see that the performance is opposite. This is because -- as stated above -- nested, inlined loops beat repeat function calls.

```js
  .add("TypedFastBitSet-nested-forof", function () {
      let card = 0;
      for (const x of tb_small) {
        for (const y of tb_small) {
          for (const z of tb_small) {
            card += x + y + z;
          }
        }
      }
      return card;
    })
    .add("TypedFastBitSet-nested-forEach", function () {
      let card = 0;
      tb_small.forEach(function (x) {
        tb_small.forEach(function (y) {
          tb_small.forEach(function (z) {
            card += x + y + z;
          });
        });
      });
      return card;
    })
```

Results:

```
TypedFastBitSet-nested-forof x 103 ops/sec ±0.40% (75 runs sampled)
TypedFastBitSet-nested-forEach x 205 ops/sec ±0.29% (87 runs sampled)
```

:) I hope this helps explain the intent behind the requested change