@benlesh 2023-05-08
Interestingly, the benchmarks you have don't show a huge difference between the naive cases from `master` to this PR... but they show a HUGE improvement when nested.

master:
```
TypedFastBitSet-forof x 97.89 ops/sec ±1.77% (71 runs sampled)
TypedFastBitSet-foreach x 186 ops/sec ±1.75% (78 runs sampled)
SparseTypedFastBitSet-forof x 100 ops/sec ±1.77% (74 runs sampled)
SparseTypedFastBitSet-foreach x 127 ops/sec ±1.82% (73 runs sampled)
TypedFastBitSet-nested-forof x 17.34 ops/sec ±2.30% (47 runs sampled)
TypedFastBitSet-nested-forEach x 203 ops/sec ±0.57% (86 runs sampled)
```

this PR:
```
TypedFastBitSet-forof x 96.65 ops/sec ±1.84% (71 runs sampled)
TypedFastBitSet-foreach x 196 ops/sec ±1.42% (83 runs sampled)
SparseTypedFastBitSet-forof x 98.03 ops/sec ±1.48% (72 runs sampled)
SparseTypedFastBitSet-foreach x 130 ops/sec ±1.76% (74 runs sampled)
TypedFastBitSet-nested-forof x 101 ops/sec ±0.55% (74 runs sampled)
TypedFastBitSet-nested-forEach x 205 ops/sec ±0.25% (87 runs sampled)
```


Anecdotally, while testing the implementation in [arbitrary Stackblitz experiments](https://stackblitz.com/edit/typescript-lvjxve?file=index.ts), I was seeing a 4x-5x improvement even in the naive case. 🤷‍♂️